net-test: packetdrill: Remove TUN_F_UFO from offload flags
Commit d591a1f3aad92ade4642e4173f4c368006c27f0f in the Linux kernel removed
handling of the TUN_F_UFO flag in the set_offload() function for tun devices (in
net/tun.c).

Because of this, when TUN_F_UFO (0x10) is set in the argument for an
ioctl(..., TUNSETOFFLOAD, ...) call on a tun device, an error will be indicated,
with errno set to EINVAL.

The output of packdrill in this case looks like this:

    TUNSETOFFLOAD: Invalid argument

This change removes the TUN_F_UFO flag from the argument in order to avoid this
problem.