packetdrill: implement relative time tolerance
packetdrill tries to accommodate difference between expected time
and observed time using a (configurable) maximum value of the difference.

Add a percentage based tolerance, which is helpful for long running tests.

Default of 0.5 % means that 5ms of drift per second is allowed.
If one event happens 10 seconds after the last one, the tolerance would be 50 ms.

Can be changed with --tolerance_percent=<percentage>