packetdrill: improve dynamic tolerance of blocking syscall
Packetdrill calculates the dynamic tolerance based on delta between time
of the last and current event. This delta value is irrelevant for a
blocking system call and the last event is unpredictable and depends
on the timing. Instead we can use the delta between the start and end
time specified in the packetdrill with an ellipse.