packetdrill: support strings including \x00 bytes
Support strings including \x00 bytes. To enable this, this commit
overhauls the string representation to represent run-time strings as a
pointer to a buffer with an accompanying length.

Note that when lexing/parsing the script, the script text is ASCII and
is not allowed to have \x00 bytes, so during lexing and parsing we can
keep things simple and still use null-terminated strings and
strlen(). It's during runtime that we unescape strings, and then can
have \x00 bytes and must be careful to use the len field and memcmp()
rather than strlen() or strcmp().

Signed-off-by: Neal Cardwell <ncardwell@google.com>