fix endianess issue on inbound DSN8/DACK8
packetdrill fails to convert the DSN8 and the DACK8 to network byte order
on little endian systems, because htole64(x) and le64toh(x) are a no-op.
On LE systems, we need to flip DSN and DACK octets before transmitting,
and leave them untouched on BE (this latter part was probably correctly
covered by be64toh() and htobe64()).

replace htonll() and ntohll() with another version that correctly flips
these octets on LE CPUs.

Signed-off-by: Davide Caratti <dcaratti@redhat.com>