packetdrill: enable sha256_32() for mp_join_syn token
The token value of mp_join_syn may be specified in the script as a
function of variable (i.e. the client or server key).  This enables
sha256_32() needed for v1 of MPTCP.

This change also fixes the calculation to retrieve the most
significant bits as specified in the RFC.

Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>