net-test: packetdrill: run_all: add --capture option
This new option is useful to capture packet traces per test for debug
purposes.

With the new parameter, we can specify to capture packets with TCPDump
and write the .pcap files in the given directory without modifying
in_netns.sh or re-launching the tests manually outside the netns.

e.g. to write .pcap files in the home directory:

  $ run_all.py -c ~

Files will be generated using the name of the test and its variant.

Note that this adds some delay: one second before starting not to miss
the first packets and one second at the end to make sure the last
packets will be captured by TCPDump.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>