net-test: packetdrill: support socket options TCP_FASTOPEN_KEY + NO_COOKIE
These two sockets were missing:

  - TCP_FASTOPEN_KEY
  - TCP_FASTOPEN_NO_COOKIE

We just need to add the new symbols to support them.

They are being added to MPTCP in the kernel and validated with
Packetdrill.

This commit was originally for the work-in-progress packetdrill fork
implementation supporting MPTCP but these options can be used for
any "plain" TCP connections.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>