net-test: packetdrill: fail test when packet doesn't arrive
In the case where a packetdrill test expects a packet it makes a
blocking recvfrom call, causing test failures due to missing packets
to hang indefinitely until the test infrastructure invoking the
packetdrill process times out. If packetdrill is being run
interactively, the user must hit ctrl-C or similar to manually
terminate the test.

This commit uses the SO_RECVTIMEO socket option to timeout the test
after the packet's expected arrival + a 2 second buffer.

If the kernel under test does not send packet within the 2-second
period, packetdrill causes the test to fail and prints an error
message like:
  error handling packet: Timed out waiting for packet

Signed-off-by: David Morley <morleyd@google.com>
Change-Id: I6999236d9fee78b73bd29ba8e82a963c8a51e2db