net-test: packetdrill: have wire_server drop incoming test traffic
Change the wire_server firewall rules so that they drop incoming test
traffic packets from the kernel under test, before they are seen by
the TCP/UDP/etc layers of the wire server machine.

In some cases (e.g., if a network does not allow spoofing) the
packetdrill test traffic may use the "real" IP addresses of the client
and server machine NICs. In such cases we don't want the kernel of the
wire server machine to process and respond to the test traffic (e.g.,
because we don't want the wire server machine's kernel to send RST
packets in response to the connections under test).

This is needed to allow wire_server tests to use the real/primary IP
addresses of the wire_client and wire_server machines, which is in
turn needed in some environments that enforce a one-to-one mapping
between ethernet switch ports and IP addresses (to prevent IP
spoofing).

This has a large side bonus: with this commit and a few other related
commits, wire_client/wire_server mode can now be used between machines
in different L2 domains, i.e. the two machines no longer need to be on
the same switch, but can be anywhere in the same L3-routable domain.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: Ifb09680cdb4fd19ea5656bce544fb0b21422e1d4