net-test: packetdrill: refine wire server packet-sniffing BPF
Stop checking ethernet src address in wire server BPF. This is
unnecessarily complicated, and doesn't work in some environments where
the switch may set its own address as the source address of ethernet
packets it transmits to hosts.

This commit carefully filters traffic so that the wire server's packet
socket only sniffs traffic that is very likely the traffic under test,
based on source and destination L4 port, etc.

This has a large side bonus: with this commit and a few other releated
commits in this patch series, wire_client/wire_server mode can now be
used between machines in different L2 domains, i.e. the two machines
no longer need to be on the same switch, but can be anywhere in the
same L3-routable domain.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: If479bc0eff18769a79786b52acaaea04a4e057dd