net-test: packetdrill: don't add/remove gateway IP if it's wire server's IP
We should not have the wire server attempt to add/remove the gateway
IP if that gateway IP is the wire server's IP. Otherwise we risk
deactivating the primary/real IP address of the wire server machine.

This new logic is appropriate for the case where the wire_client and
wire_server machines are using their real/primary IP addresses for the
test traffic.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: I1778e69df2ee1494132e2f1938d50056ac7da125