net-test: packetdrill: useful defaults for --wire_client_dev, --wire_server_dev, --ip_version
To make life easier for users, implement useful defaults for
--wire_client_dev, --wire_server_dev, --ip_version, so that in the
common case when using wire_server mode users don't have to specify
any of these parameters.

For --wire_client_dev and --wire_server_dev use the name of the
primary NIC, estimated as the NIC that has the IP address to which the
machine's hostname resolves (using string_to_ip(), which uses
getaddrinfo(3)).

If we can't determine this NIC, then print a useful error message and
exit with an error. This seems preferable to the previous approach of
assuming "eth0", which used to be a common ethernet device name on
Linux but is not a common ethernet device name on the most popular
Linux OS distribution now.

For --ip_version, default to the version matching the IP address
family of the machine's primary IP. If we can't determine that, use
the prvious default of IPv4.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: I7e3bd717edf2c3e05b9e8620addc3e9429f984f5