net-test: packetdrill: have wire_client send live_local_ip_string to wire_server
So that the wire_server knows the destination IP address to use for
packets it injects into the network, have the wire_client send this
information to the wire_server before the test begins.

Previously this was not necessary. But, for convenience, an earlier
patch in this series allows the wire_client to set the default
live_local_ip_string implicitly using the dns_hostname->IP mapping
from the client machine. So now the wire_client needs to send this
default live_local_ip_string information to the wire_server. The
wire_server may not have access to the dns_hostname->IP mapping that
the client has (e.g., the dns_hostname->IP mapping may have come from
/etc/hosts on the client machine).

Communicating this live_local_ip_string information explicitly allows
the wire_server to start with the same default live_local_ip_string as
the wire_client, so that after processing all the command line
arguments the wire_server ends up with the same configuration as the
wire_client.

Signed-off-by: Neal Cardwell <ncardwell@google.com>
Change-Id: I2781a64f1d200ed31b439c7e12629352db2e6f75