net-test: packetdrill: bug fix: wire_server must ignore new --wire_server_at arg
Fix a bug reported on the packetdrill mailing list:
  https://groups.google.com/g/packetdrill/c/SIZSjTe8uTM/m/u56xAVQTAAAJ

The debug output in packetdrill_server.debug.output showed that the
wire_server was parsing the --wire_server_at=10.4.99.103 argument,
presumably causing the wire_server to set config->is_wire_client =
true while parsing that argument, thus making the wire_server believe
it was a wire_client, and thus causing it to die with the error
message:

  please specify --wire_client_dev=<eth_dev_name>

The fix is to ensure that packetdrill treats the new --wire_server_at
argument like the existing --wire_client argument, where we ensure
that the wire server does not use this argument.

While we're at it, tighten up the code to ignore client-only args, and
enforce this on the server side in wire_server_unserialize_argv(), in
case the client-sdie wire_client_serialize_argv() is buggy (as it was
here).

Signed-off-by: Neal Cardwell <ncardwell@google.com>