net-test: packetdrill: fix the icmp echo bytes
When a icmp packet use opt_icmp_echoed option, the
payload_bytes which provided by UDP, is not add
in icmp_bytes yet, this would make paylaod error when
packet verify. For example:

in script the packet event like this:
	...
	+0 > icmp unreachable port_unreachable[udp(20)]
It will result in the following errors:
	live packet payload: expected 28 bytes vs actual 48 bytes

The packet_payload_len function is not contains the payload_bytes.
Add the payload_bytes in echoed_bytes to fix this error.

Signed-off-by: Rongguang Wei <weirongguang@kylinos.cn>